Create 2) A Simple Class for Complex Numbers
Complex numbers are an important mathematical concept that extends the real numbers by bringing the imaginary unit 
i
 into the picture. Your task is to write a class ComplexNumber that represents a complex number and offers methods to calculate with complex numbers.

Mathematically, a complex number has the form 
c
=
a
+
b
i
, where 
a
 and 
b
 are real numbers (elements of 
R
) and 
i
 satisfies the equation 
i
2
=
−
1
. 
a
 is referred to as the real part of 
c
 and 
b
 is the imaginary part of 
c
. The set of complex numbers is denoted 
C
 and is a strict superset of 
R
; you can think of a real number as a complex number where the imaginary part 
b
i
 is 0. For a further discussion of complex numbers and how the mathematical operations on them are defined see the Wikipedia entry on complex numbers.

This is what your class ComplexNumber should provide:

Two private fields called re and im that represent the real and the imaginary part of your complex number.
Three public constructors. ComplexNumber() initializes the real and imaginary parts with 0. ComplexNumber(double real, double imaginary) initializes the real and imaginary part with the given values. ComplexNumber(ComplexNumber cn) initializes the real and imaginary parts with the values from the given complex number (effectively “copying” a complex number object).
Getters to access the real and imaginary part of your complex number, called getReal() and getImaginary(). Your class should not provide setters. Once a complex number is instantiated, it cannot be changed anymore.
Methods to calculate stuff, all of which return a new complex number instead of changing the current one.
conjugate() returns a new complex number which is the conjugate of the one the methods was called on.
add(ComplexNumber other) returns a new complex number which is the sum of the two.
subtract(ComplexNumber other) returns a new complex number which is the result of subtracting other from the current complex number.
multiply(ComplexNumber other) returns a new complex number which is the product of the two.
divide(ComplexNumber other) returns null if the real and imaginary part of other is 0. Otherwise, it returns a new complex number which is the result of dividing the current complex number by the other one.
abs() returns a double that is the absolute value of the complex number. (Note: The Math class provides a method to calculate square roots.)
A toString() method that returns a sensible string representation of the complex number. A complex number with real part 42 and imaginary part 23 could for instance be written as 42.0 + 23.0i.
Remember to add complete Javadoc comments to your class as well as all public constructors and methods (adding them to your private fields won’t hurt either). Be sure to document parameters and return values as well.

This assignment will help you

practice writing your own classes and methods,
learn how to write proper Javadoc comments, and
familiarize yourself with complex numbers.