Create Time Zones
Some think madness lies in learning to program. We think madness much rather lies in learning to program things that have something to do with time zones. Such as a class that represents a date, as in this assignment.

Write a class called Date, placed in a package called programming.set6.date and provide the following public entries:

A static method getDaysInMonth(int year, int month) that returns the number of days in the given month of the given year, where 1 represents January and 12 represents December. If month is out of range, return 0.
A static method validate(int year, int month, int day) that checks if the given three integers represent a valid date. Remember that the months differ in how many days they have. (Think about how you can easily represent the number of days each month has.)
A constructor that takes the year, the month, and the day to initialize the Date instance, each as an integer. The constructor should check if the three numbers specifiy a valid date. If they do not, throw an IllegalArgumentException which indicates that the date is not valid. Throwing an exception will immediately terminate the constructor call and your program. Here is the code for throwing that exception:
throw new IllegalArgumentException("This is not a valid date.");
Getter methods getDay(), getMonth(), and getYear().
A method dayOfYear() that returns the number of the day represented by the Date instance. For example, January 20th is the 20th day of the year, while May 20th is the 140th day of the year.
A method sameYearDiff(Date other) that calculates how many days the other Date is off and returns that as an integer. If other lies in the future, that number is positive; otherwise, it is negative. The method is only properly defined if both Date instances represent dates in the same year. Otherwise, return 0.
A toString() method that returns the date in String form like this: November 20, 2015.
Remember that February has 29 days in leap years. A year is a leap year if it is divisible by 4 and not divisible by 100 (unless it is also divisible by 400). We expect you to write complete Javadoc comments for your class as well as for all its entries.

Doing this assignment will help you practice designing classes as well as control flow statements. There is also an opportunity for using arrays.

Note: The Java runtime library of course already provides classes that can handle dates. However, we want you to write your own class, without using any of the predefined ones. We want you to learn stuff.