Create Sudoku
While the title of this assignment sounds impressively important and complicated, the actual assignment is to implement a class that models a Sudoku board. That sound you’re hearing are your hopes and dreams for this assignment being crushed. You just learned how life works.

Wikipedia explains Sudoku like this:

“The objective is to fill a 9×9 grid with digits so that each column, each row, and each of the nine 3×3 sub-grids that compose the grid (also called “boxes”, “blocks”, “regions”, or “sub-squares”) contains all of the digits from 1 to 9. The puzzle setter provides a partially completed grid, which for a well-posed puzzle has a unique solution.”

We have written a class called NumberBoard that you can download here. Do so, and import it into your Eclipse workspace. You will find that the class models a board that consists of rows and columns, thus forming cells. Each cell can either be empty or be filled with an arbitrary integer 
≥
0
. Read the Javadoc comments for all of the class’s entries to familiarize yourself with the functionality the class provides. During this assignment, leave this class unchanged.

Create a subclass of NumberBoard called Sudoku in the same package as the NumberBoard class. Implement the following features:

NumberBoard has a constructor that can construct number boards of arbitrary sizes. However, Sudoku boards always have nine rows and nine columns. Add a constructor to Sudoku that doesn’t take any parameters and invokes the superclass constructor in a way to construct a board with nine rows and nine columns.
The implementation of the setValueAt(int row, int col, int value) method in NumberBoard checks if value is valid before writing it into the specified cell. As far as NumberBoard is concerned, value is valid if it is either NumberBoard.EMPTY or 
≥
0
. However, in a Sudoku board, value can either be empty or it must hold that 
1
≤
v
a
l
u
e
≤
9
. Overwrite setValueAt(...) such that the superclass implementation is only called if value is valid in the Sudoku sense. Note that you shouldn’t check if the Sudoku board overall is valid in this method. That’s what the next method does.
Speaking of the next method: implement a method isValid() that returns true if the Sudoku board’s current state is valid and false if it is not. The state is valid if there is no row, no column, and no 3×3 sub-grid that contains a number more than once. Note that there can be arbitrarily many empty fields. You may want to write a private method each to check the rows, columns, and sub-grids for validity, but that is up to you.
Override the toString() method such that it returns a String that represents the current state of the board. Each row should be represented by a line in the string. Each pair of cells in a row should be divided by a space character. Empty cells should be represented by a space character. You can start a new line in the string by adding the newline character, "\n". Printing the result to the console should look like this: