Create Hashmaps
Completing this assignment will help you understand the concept of hash maps and bucket hashing.

Maps as a data structure are arguably way more useful than Apple Maps was at its inception. Which is why instead of asking you to implement your own Maps app, we’re asking you to implement your own map data structure.

Lecture 11 taught you about the basics of Java’s HashMap class. You even implemented a simple map based on bucket hashing for strings. In this assignment, we want you to take those ideas and extend them into a map implementation that can be used with arbitrary types for keys and values. Since there already is a HashMap class, your class will be called BrownieMap.

Create the class in a package called programming.set11.brownies. It should be a generic class that accepts two type parameters:

K is the type of the map’s keys.
V is the type of the map’s values.